Use Post/Redirect/Get pattern for posting words.
When posting new words, we now post to the /add end-point, which adds
the words to the database, stores state in the session (i.e. cookies)
and redirects to the original page. This way the original page can be
safely bookmarked or refreshed even after post.

We now need a secret key configured for signing cookies. This is read
from the SECRET_KEY environment variable. When running using
docker-compose, the value should be set in a .env file. An example
.env.example file is included.