Fix "remove word" button in word list page.
The word list page lacked a message box to show the success/failure
message when a word was removed. Also, the remove code would attempt
to update the word count, which is only available on the web page, and
fail. The code now checks to make sure the word count label actually
exist.

Removal of a word, would now remove the who li element, instead of the
span element, so a bit extra empty space is not left when a word is
removed.