Show message in separate line in list page.
Because of the "display: inline-block" style of the "msgbox" class,
the success/failure message would be shown in the same line as the
word list. By wrapping the word list in a div, we now make sure the
message is shown in a separate line.

The "display: inline-block" is necessary so that the message box is
not expanded to the whole screen.